+++ steinbock/_env.py
@@ -7,7 +7,9 @@ from functools import wraps
 from pathlib import Path
 
 
-def run_captured(args, *popen_args, file=sys.stdout, **popen_kwargs):
+def run_captured(
+    args, *popen_args, file=sys.stdout, **popen_kwargs
+) -> subprocess.CompletedProcess:
     with subprocess.Popen(
         args,
         *popen_args,
@@ -55,6 +57,7 @@ def use_ilastik_env(func):
             kwargs["ilastik_env"] = os.environ.copy()
         kwargs["ilastik_env"].pop("PYTHONPATH", None)
         kwargs["ilastik_env"].pop("PYTHONHOME", None)
+        kwargs["ilastik_env"].pop("LD_LIBRARY_PATH", None)
         return func(*args, **kwargs)
 
     return use_ilastik_env_wrapper