+++ steinbock/classification/_cli/ilastik.py
@@ -133,10 +133,10 @@ def prepare_cmd(
         mean_factor=mean_factor,
         scale_factor=scale_factor,
     ):
-        ilastik_img_stem = Path(ilastik_img_dir) / f"{img_file.stem}"
-        ilastik_img_file = ilastik.write_ilastik_image(
-            ilastik_img, ilastik_img_stem
+        ilastik_img_file = io._as_path_with_suffix(
+            Path(ilastik_img_dir) / img_file.name, ".h5"
         )
+        ilastik.write_ilastik_image(ilastik_img, ilastik_img_file)
         ilastik_img_files.append(ilastik_img_file)
         click.echo(ilastik_img_file)
         del ilastik_img
@@ -151,14 +151,12 @@ def prepare_cmd(
         ilastik_img_files, ilastik_crop_size, np.random.default_rng(seed=seed)
     ):
         if ilastik_crop is not None:
-            ilastik_crop_stem = Path(ilastik_crop_dir) / (
+            ilastik_crop_file = Path(ilastik_crop_dir) / (
                 f"{ilastik_img_file.stem}"
                 f"_x{ilastik_crop_x}_y{ilastik_crop_y}"
-                f"_w{ilastik_crop_size}_h{ilastik_crop_size}"
+                f"_w{ilastik_crop_size}_h{ilastik_crop_size}.h5"
             )
-            ilastik_crop_file = ilastik.write_ilastik_crop(
-                ilastik_crop, ilastik_crop_stem
-            )
+            ilastik.write_ilastik_crop(ilastik_crop, ilastik_crop_file)
             ilastik_crop_files.append(ilastik_crop_file)
             click.echo(ilastik_crop_file)
             del ilastik_crop
@@ -170,6 +168,7 @@ def prepare_cmd(
     ilastik.create_and_save_ilastik_project(
         ilastik_crop_files, ilastik_project_file
     )
+    click.echo(ilastik_project_file)
 
 
 @ilastik_cmd_group.command(
@@ -296,16 +295,18 @@ def fix_cmd(
             Path(ilastik_project_file).name + ".bak"
         )
         if ilastik_project_backup_file.exists():
-            return click.echo(
+            click.echo(
                 "ERROR: Ilastik project backup file exists", file=sys.stderr
             )
+            return
         ilastik_crop_backup_dir = Path(ilastik_crop_dir).with_name(
             Path(ilastik_crop_dir).name + ".bak"
         )
         if ilastik_crop_backup_dir.exists():
-            return click.echo(
+            click.echo(
                 "ERROR: Ilastik crop backup directory exists", file=sys.stderr
             )
+            return
         shutil.copyfile(ilastik_project_file, ilastik_project_backup_file)
         ilastik_crop_backup_dir.mkdir()
         for ilastik_crop_file in ilastik_crop_files:
@@ -323,12 +324,11 @@ def fix_cmd(
         ilastik_crop_files, orig_axis_order=orig_axis_order
     ):
         if last_transpose_axes not in (None, transpose_axes):
-            return click.echo(
+            click.echo(
                 "ERROR: Inconsistent axis orders across crops", file=sys.stderr
             )
-        ilastik_crop_file = ilastik.write_ilastik_crop(
-            ilastik_crop, ilastik_crop_file
-        )
+            return
+        ilastik.write_ilastik_crop(ilastik_crop, ilastik_crop_file)
         ilastik_crop_shapes[ilastik_crop_file.stem] = ilastik_crop.shape
         last_transpose_axes = transpose_axes
         click.echo(ilastik_crop_file)
@@ -340,3 +340,4 @@ def fix_cmd(
         ilastik_crop_shapes,
         last_transpose_axes,
     )
+    click.echo(ilastik_project_file)