+++ steinbock/export/_cli/data.py
@@ -32,12 +32,14 @@ def csv_cmd(data_dirs, concatenate, csv_file_or_dir):
     if not data_dirs:
         return  # empty variadic argument, gracefully degrade into noop
     if concatenate and Path(csv_file_or_dir).is_dir():
-        return click.echo(
+        click.echo(
             "ERROR: Specify a single output file when concatenating",
             file=sys.stderr,
         )
+        return
     elif not concatenate and Path(csv_file_or_dir).is_file():
-        return click.echo("ERROR: Output directory is a file", file=sys.stderr)
+        click.echo("ERROR: Output directory is a file", file=sys.stderr)
+        return
     first_data_files = io.list_data_files(data_dirs[0])
     data_file_lists = [first_data_files]
     for data_dir in data_dirs[1:]:
@@ -55,14 +57,15 @@ def csv_cmd(data_dirs, concatenate, csv_file_or_dir):
                 df.to_csv(csv_file_or_dir, index=False)
             else:
                 df.to_csv(csv_file_or_dir, header=False, index=False, mode="a")
-            click.echo(img_file_name)
         else:
             csv_file = io._as_path_with_suffix(
                 Path(csv_file_or_dir) / img_file_name, ".csv"
             )
             df.to_csv(csv_file, index=False)
-            click.echo(csv_file.name)
+            click.echo(csv_file)
         del df
+    if concatenate:
+        click.echo(csv_file_or_dir)
 
 
 @click.command(name="fcs", help="Merge and export object data to FCS")
@@ -87,12 +90,14 @@ def fcs_cmd(data_dirs, concatenate, fcs_file_or_dir):
     if not data_dirs:
         return  # empty variadic argument, gracefully degrade into noop
     if concatenate and Path(fcs_file_or_dir).is_dir():
-        return click.echo(
+        click.echo(
             "ERROR: Specify a single output file when concatenating",
             file=sys.stderr,
         )
+        return
     elif not concatenate and Path(fcs_file_or_dir).is_file():
-        return click.echo("ERROR: Output directory is a file", file=sys.stderr)
+        click.echo("ERROR: Output directory is a file", file=sys.stderr)
+        return
     first_data_files = io.list_data_files(data_dirs[0])
     data_file_lists = [first_data_files]
     for data_dir in data_dirs[1:]:
@@ -112,18 +117,18 @@ def fcs_cmd(data_dirs, concatenate, fcs_file_or_dir):
         df,
     ) in data.try_convert_to_dataframe_from_disk(*data_file_lists):
         if concatenate:
-            click.echo(img_file_name)
             dfs.append(df)
         else:
             fcs_file = io._as_path_with_suffix(
                 Path(fcs_file_or_dir) / img_file_name, ".fcs"
             )
             write_fcs(fcs_file, df.columns.values, df.values)
-            click.echo(fcs_file.name)
+            click.echo(fcs_file)
             del df
     if concatenate:
         df = pd.concat(dfs, ignore_index=True, copy=False)
         write_fcs(fcs_file_or_dir, df.columns.values, df.values)
+        click.echo(fcs_file_or_dir)
 
 
 @click.command(name="anndata", help="Merge and export object data to AnnData")
@@ -197,24 +202,24 @@ def anndata_cmd(
 ):
     if anndata_format == "h5ad":
 
-        def write_anndata(adata, anndata_file, keep_suffix=False):
-            suffix = Path(anndata_file).suffix if keep_suffix else ".h5ad"
+        def write_anndata(adata, anndata_file, ignore_suffix=False):
+            suffix = Path(anndata_file).suffix if ignore_suffix else ".h5ad"
             anndata_file = io._as_path_with_suffix(anndata_file, suffix)
             adata.write_h5ad(anndata_file)
             return anndata_file
 
     elif anndata_format == "loom":
 
-        def write_anndata(adata, anndata_file, keep_suffix=False):
-            suffix = Path(anndata_file).suffix if keep_suffix else ".loom"
+        def write_anndata(adata, anndata_file, ignore_suffix=False):
+            suffix = Path(anndata_file).suffix if ignore_suffix else ".loom"
             anndata_file = io._as_path_with_suffix(anndata_file, suffix)
             adata.write_loom(anndata_file)
             return anndata_file
 
     elif anndata_format == "zarr":
 
-        def write_anndata(adata, anndata_file, keep_suffix=False):
-            suffix = Path(anndata_file).suffix if keep_suffix else ".zarr"
+        def write_anndata(adata, anndata_file, ignore_suffix=False):
+            suffix = Path(anndata_file).suffix if ignore_suffix else ".zarr"
             anndata_file = io._as_path_with_suffix(anndata_file, suffix)
             adata.write_zarr(anndata_file)
             return anndata_file
@@ -240,12 +245,14 @@ def anndata_cmd(
     # if image_info_file is not None and Path(image_info_file).is_file():
     #     image_info = io.read_image_info(image_info_file)
     if concatenate and Path(anndata_file_or_dir).is_dir():
-        return click.echo(
+        click.echo(
             "ERROR: Specify a single output file when concatenating",
             file=sys.stderr,
         )
+        return
     elif not concatenate and Path(anndata_file_or_dir).is_file():
-        return click.echo("ERROR: Output directory is a file", file=sys.stderr)
+        click.echo("ERROR: Output directory is a file", file=sys.stderr)
+        return
     if concatenate:
         click.echo(
             "WARNING: The anndata package currently does not support on-disk; "
@@ -269,13 +276,12 @@ def anndata_cmd(
     ):
         if concatenate:
             adatas[img_file_name] = adata
-            click.echo(img_file_name)
         else:
             anndata_file = io._as_path_with_suffix(
                 Path(anndata_file_or_dir) / img_file_name, ".adata"
             )
             anndata_file = write_anndata(adata, anndata_file)
-            click.echo(anndata_file.name)
+            click.echo(anndata_file)
             del adata
     if concatenate:
         adata = anndata_concat(
@@ -284,4 +290,5 @@ def anndata_cmd(
         obs_cols = list(adata.obs.columns)
         obs_cols.insert(0, obs_cols.pop(obs_cols.index("Image")))
         adata.obs = adata.obs.loc[:, obs_cols]
-        write_anndata(adata, anndata_file_or_dir, keep_suffix=True)
+        write_anndata(adata, anndata_file_or_dir, ignore_suffix=True)
+        click.echo(anndata_file_or_dir)