+++ steinbock/io.py
@@ -2,6 +2,7 @@ import imageio
 import numpy as np
 import os
 import pandas as pd
+import re
 import tifffile
 
 from os import PathLike
@@ -13,12 +14,15 @@ img_dtype = np.dtype(os.environ.get("STEINBOCK_IMG_DTYPE", "float32"))
 mask_dtype = np.dtype(os.environ.get("STEINBOCK_MASK_DTYPE", "uint16"))
 
 
-def _as_path_with_suffix(
-    path: Union[str, PathLike], suffix: str, replace_ome_suffix: bool = True
-) -> Path:
+def _as_path_with_suffix(path: Union[str, PathLike], suffix: str) -> Path:
     path = Path(path)
-    if replace_ome_suffix and path.name.endswith(".ome.tiff"):
-        path = path.with_name(path.name[:-9] + ".tiff")
+    if re.fullmatch(r".+\.ome\.[^.]+", path.name, flags=re.IGNORECASE):
+        stem, ome_suffix, suffix = path.name.rpartition(".ome.")
+        if ome_suffix:
+            path = path.with_name(f"{stem}.{suffix}")
+        stem, ome_suffix, suffix = path.name.rpartition(".OME.")
+        if ome_suffix:
+            path = path.with_name(f"{stem}.{suffix}")
     return path.with_suffix(suffix)
 
 
@@ -59,9 +63,8 @@ def _list_related_files(
 
 
 def read_panel(
-    panel_stem: Union[str, PathLike], kept_only: bool = True
+    panel_file: Union[str, PathLike], unfiltered: bool = False
 ) -> pd.DataFrame:
-    panel_file = _as_path_with_suffix(panel_stem, ".csv")
     panel = pd.read_csv(
         panel_file,
         sep=",|;",
@@ -90,19 +93,17 @@ def read_panel(
                 raise ValueError(
                     f"Duplicated values for '{unique_col}' in {panel_file}"
                 )
-    if kept_only and "keep" in panel:
+    if not unfiltered and "keep" in panel:
         panel = panel.loc[panel["keep"].astype(bool), :]
     return panel
 
 
-def write_panel(panel: pd.DataFrame, panel_stem: Union[str, PathLike]) -> Path:
-    panel_file = _as_path_with_suffix(panel_stem, ".csv")
+def write_panel(panel: pd.DataFrame, panel_file: Union[str, PathLike]) -> None:
     panel = panel.copy()
     for col in panel.columns:
         if panel[col].convert_dtypes().dtype == pd.BooleanDtype():
             panel[col] = panel[col].astype(pd.UInt8Dtype())
     panel.to_csv(panel_file, index=False)
-    return panel_file
 
 
 def list_image_files(
@@ -116,43 +117,64 @@ def list_image_files(
 
 def read_image(
     img_file: Union[str, PathLike],
-    keep_suffix: bool = False,
     use_imageio: bool = False,
     native_dtype: bool = False,
 ) -> np.ndarray:
-    if not keep_suffix:
-        img_file = _as_path_with_suffix(
-            img_file, ".tiff", replace_ome_suffix=False
-        )
     if use_imageio:
         img = imageio.volread(img_file)
+        orig_img_shape = img.shape
+        while img.ndim > 3 and img.shape[0] == 1:
+            img = img.sqeeze(axis=0)
+        while img.ndim > 3 and img.shape[-1] == 1:
+            img = img.sqeeze(axis=img.ndim - 1)
+        if img.ndim == 2:
+            img = img[np.newaxis, :, :]
+        elif img.ndim != 3:
+            raise ValueError(
+                f"Unsupported shape {orig_img_shape} for image {img_file}"
+            )
     else:
-        img = tifffile.imread(img_file)
-    while img.ndim > 3 and img.shape[0] == 1:
-        img = img.sqeeze(axis=0)
-    while img.ndim > 3 and img.shape[-1] == 1:
-        img = img.sqeeze(axis=img.ndim - 1)
-    if img.ndim == 2:
-        img = img[np.newaxis, :, :]
-    elif img.ndim != 3:
-        raise ValueError(f"Unsupported number of image dimensions: {img_file}")
+        img = tifffile.imread(img_file, squeeze=False)
+        if img.ndim == 2:
+            img = img[np.newaxis, :, :]
+        elif img.ndim == 5:
+            size_t, size_z, size_c, size_y, size_x = img.shape
+            if size_t != 1 or size_z != 1:
+                raise ValueError(
+                    f"{img_file}: unsupported TZCYX shape {img.shape}"
+                )
+            img = img[0, 0, :, :, :]
+        elif img.ndim == 6:
+            size_t, size_z, size_c, size_y, size_x, size_s = img.shape
+            if size_t != 1 or size_z != 1 or size_s != 1:
+                raise ValueError(
+                    f"{img_file}: unsupported TZCYXS shape {img.shape}"
+                )
+            img = img[0, 0, :, :, :, 0]
+        elif img.ndim != 3:
+            raise ValueError(
+                f"{img_file}: unsupported number of dimensions ({img.ndim})"
+            )
     if not native_dtype:
         img = _to_dtype(img, img_dtype)
     return img
 
 
 def write_image(
-    img: np.ndarray, img_stem: Union[str, PathLike], ignore_dtype: bool = False
-) -> Path:
+    img: np.ndarray,
+    img_file: Union[str, PathLike],
+    ignore_dtype: bool = False,
+) -> None:
     if not ignore_dtype:
         img = _to_dtype(img, img_dtype)
-    img_file = _as_path_with_suffix(img_stem, ".tiff")
-    tifffile.imwrite(img_file, data=img, imagej=True)
-    return img_file
+    tifffile.imwrite(
+        img_file,
+        data=img[np.newaxis, np.newaxis, :, :, :, np.newaxis],
+        imagej=True,
+    )
 
 
-def read_image_info(image_info_stem: Union[str, PathLike]) -> pd.DataFrame:
-    image_info_file = _as_path_with_suffix(image_info_stem, ".csv")
+def read_image_info(image_info_file: Union[str, PathLike]) -> pd.DataFrame:
     image_info = pd.read_csv(
         image_info_file,
         sep=",|;",
@@ -185,11 +207,9 @@ def read_image_info(image_info_stem: Union[str, PathLike]) -> pd.DataFrame:
 
 
 def write_image_info(
-    image_info: pd.DataFrame, image_info_stem: Union[str, PathLike]
-) -> Path:
-    image_info_file = _as_path_with_suffix(image_info_stem, ".csv")
+    image_info: pd.DataFrame, image_info_file: Union[str, PathLike]
+) -> None:
     image_info.to_csv(image_info_file, index=False)
-    return image_info_file
 
 
 def list_mask_files(
@@ -201,23 +221,40 @@ def list_mask_files(
     return sorted(Path(mask_dir).rglob("*.tiff"))
 
 
-def read_mask(mask_stem: Union[str, PathLike]) -> np.ndarray:
-    mask_file = _as_path_with_suffix(mask_stem, ".tiff")
-    mask = tifffile.imread(mask_file)
-    while mask.ndim > 2 and mask.shape[0] == 1:
-        mask = mask.sqeeze(axis=0)
-    while mask.ndim > 2 and mask.shape[-1] == 1:
-        mask = mask.sqeeze(axis=mask.ndim - 1)
-    if mask.ndim != 2:
-        raise ValueError(f"Unsupported number of mask dimensions: {mask_file}")
-    return _to_dtype(mask, mask_dtype)
+def read_mask(
+    mask_file: Union[str, PathLike], native_dtype: bool = False
+) -> np.ndarray:
+    mask = tifffile.imread(mask_file, squeeze=False)
+    if mask.ndim == 5:
+        size_t, size_z, size_c, size_y, size_x = mask.shape
+        if size_t != 1 or size_z != 1 or size_c != 1:
+            raise ValueError(
+                f"{mask_file}: unsupported TZCYX shape {mask.shape}"
+            )
+        mask = mask[0, 0, 0, :, :]
+    elif mask.ndim == 6:
+        size_t, size_z, size_c, size_y, size_x, size_s = mask.shape
+        if size_t != 1 or size_z != 1 or size_c != 1 or size_s != 1:
+            raise ValueError(
+                f"{mask_file}: unsupported TZCYXS shape {mask.shape}"
+            )
+        mask = mask[0, 0, 0, :, :, 0]
+    elif mask.ndim != 2:
+        raise ValueError(
+            f"{mask_file}: unsupported number of dimensions ({mask.ndim})"
+        )
+    if not native_dtype:
+        mask = _to_dtype(mask, mask_dtype)
+    return mask
 
 
-def write_mask(mask: np.ndarray, mask_stem: Union[str, PathLike]) -> Path:
+def write_mask(mask: np.ndarray, mask_file: Union[str, PathLike]) -> None:
     mask = _to_dtype(mask, mask_dtype)
-    mask_file = _as_path_with_suffix(mask_stem, ".tiff")
-    tifffile.imwrite(mask_file, data=mask, imagej=True)
-    return mask_file
+    tifffile.imwrite(
+        mask_file,
+        data=mask[np.newaxis, np.newaxis, np.newaxis, :, :, np.newaxis],
+        imagej=True,
+    )
 
 
 def list_data_files(
@@ -229,18 +266,15 @@ def list_data_files(
     return sorted(Path(data_dir).rglob("*.csv"))
 
 
-def read_data(data_stem: Union[str, PathLike]) -> pd.DataFrame:
-    data_file = _as_path_with_suffix(data_stem, ".csv")
+def read_data(data_file: Union[str, PathLike]) -> pd.DataFrame:
     return pd.read_csv(
         data_file, sep=",|;", index_col="Object", engine="python"
     )
 
 
-def write_data(data: pd.DataFrame, data_stem: Union[str, PathLike]) -> Path:
-    data_file = _as_path_with_suffix(data_stem, ".csv")
+def write_data(data: pd.DataFrame, data_file: Union[str, PathLike]) -> None:
     data = data.reset_index()
     data.to_csv(data_file, index=False)
-    return data_file
 
 
 def list_neighbors_files(
@@ -252,8 +286,7 @@ def list_neighbors_files(
     return sorted(Path(neighbors_dir).rglob("*.csv"))
 
 
-def read_neighbors(neighbors_stem: Union[str, PathLike]) -> pd.DataFrame:
-    neighbors_file = _as_path_with_suffix(neighbors_stem, ".csv")
+def read_neighbors(neighbors_file: Union[str, PathLike]) -> pd.DataFrame:
     return pd.read_csv(
         neighbors_file,
         sep=",|;",
@@ -268,9 +301,8 @@ def read_neighbors(neighbors_stem: Union[str, PathLike]) -> pd.DataFrame:
 
 
 def write_neighbors(
-    neighbors: pd.DataFrame, neighbors_stem: Union[str, PathLike]
-) -> Path:
-    neighbors_file = _as_path_with_suffix(neighbors_stem, ".csv")
+    neighbors: pd.DataFrame, neighbors_file: Union[str, PathLike]
+) -> None:
     neighbors = neighbors.loc[:, ["Object", "Neighbor", "Distance"]].astype(
         {
             "Object": mask_dtype,
@@ -279,4 +311,3 @@ def write_neighbors(
         }
     )
     neighbors.to_csv(neighbors_file, index=False)
-    return neighbors_file