+++ steinbock/preprocessing/_cli/imc.py
@@ -3,8 +3,10 @@ import pandas as pd
 import sys
 
 from contextlib import nullcontext
+from os import PathLike
 from pathlib import Path
 from tempfile import TemporaryDirectory
+from typing import List, Union
 from zipfile import ZipFile
 
 from steinbock import io
@@ -119,10 +121,12 @@ def panel_cmd(
         if len(txt_files) > 0:
             panel = imc.create_panel_from_txt_files(txt_files)
     if panel is None:
-        return click.echo(
+        click.echo(
             "ERROR: No panel/.mcd/.txt file found", file=sys.stderr
         )
+        return
     io.write_panel(panel, panel_file)
+    click.echo(panel_file)
 
 
 @imc_cmd_group.command(
@@ -211,7 +215,7 @@ def images_cmd(
             if img_file_stem in mcd_txt_files:
                 num_dupl += 1
                 first_mcd_txt_file = mcd_txt_files[img_file_stem][0]
-                img_file_stem = f"DUPLICATE{num_dupl:03d}_" + img_file_stem
+                img_file_stem = f"DUPLICATE{num_dupl:03d}_{img_file_stem}"
                 click.echo(
                     f"WARNING: File {mcd_txt_file} is a duplicate of "
                     f"{first_mcd_txt_file}, saving as {img_file_stem}",
@@ -220,7 +224,8 @@ def images_cmd(
                 mcd_txt_files[img_file_stem].append(mcd_txt_file)
             else:
                 mcd_txt_files[img_file_stem] = [mcd_txt_file]
-            img_file = io.write_image(img, Path(img_dir) / img_file_stem)
+            img_file = Path(img_dir) / f"{img_file_stem}.tiff"
+            io.write_image(img, img_file)
             recovery_file_name = None
             if recovery_file is not None:
                 recovery_file_name = recovery_file.name
@@ -259,46 +264,38 @@ def images_cmd(
             del img
     image_info = pd.DataFrame(data=image_info_data)
     io.write_image_info(image_info, image_info_file)
+    click.echo(image_info_file)
 
 
-def _extract_zips(path, suffix, dest):
+def _extract_zips(
+    path: Union[str, PathLike], suffix: str, dest: Union[str, PathLike]
+) -> List[Path]:
     extracted_files = []
     for zip_file_path in Path(path).rglob("*.zip"):
         with ZipFile(zip_file_path) as zip_file:
             zip_infos = sorted(zip_file.infolist(), key=lambda x: x.filename)
             for zip_info in zip_infos:
-                zip_info_suffix = Path(zip_info.filename).suffix
-                if not zip_info.is_dir() and zip_info_suffix == suffix:
+                if not zip_info.is_dir() and zip_info.filename.endswith(
+                    suffix
+                ):
                     extracted_file = zip_file.extract(zip_info, path=dest)
                     extracted_files.append(Path(extracted_file))
     return extracted_files
 
 
-def _collect_mcd_files(mcd_dir, unzip_dir=None):
+def _collect_mcd_files(
+    mcd_dir: Union[str, PathLike], unzip_dir: bool = None
+) -> List[Path]:
     mcd_files = imc.list_mcd_files(mcd_dir)
     if unzip_dir is not None:
         mcd_files += _extract_zips(mcd_dir, ".mcd", unzip_dir)
-    # unique_mcd_stem_names = []
-    # for mcd_file in mcd_files:
-    #     if mcd_file.stem in unique_mcd_stem_names:
-    #         return click.echo(
-    #             f"ERROR: Duplicated file stem {mcd_file.stem}",
-    #             file=sys.stderr
-    #         )
-    #     unique_mcd_stem_names.append(mcd_file.stem)
     return mcd_files
 
 
-def _collect_txt_files(txt_dir, unzip_dir=None):
+def _collect_txt_files(
+    txt_dir: Union[str, PathLike], unzip_dir: bool = None
+) -> List[Path]:
     txt_files = imc.list_txt_files(txt_dir)
     if unzip_dir is not None:
         txt_files += _extract_zips(txt_dir, ".txt", unzip_dir)
-    # unique_txt_stem_names = []
-    # for txt_file in txt_files:
-    #     if txt_file.stem in unique_txt_stem_names:
-    #         return click.echo(
-    #             f"ERROR: Duplicated file stem {txt_file.stem}",
-    #             file=sys.stderr
-    #         )
-    #     unique_txt_stem_names.append(txt_file.stem)
     return txt_files