+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.12.0] - 2022-01-08
+
+  - In-memory mosaics
+  - Stabilized image/mask IO
+  - Standardized command output
+  - Minor changes, documentation
+
 ## [0.11.0] - 2021-12-28
 
   - Removed steinbock version check
@@ -272,6 +279,7 @@ Added:
 Initial release for beta testing
 
 
+[0.12.0]: https://github.com/BodenmillerGroup/steinbock/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/BodenmillerGroup/steinbock/compare/v0.10.7...v0.11.0
 [0.10.7]: https://github.com/BodenmillerGroup/steinbock/compare/v0.10.6...v0.10.7
 [0.10.6]: https://github.com/BodenmillerGroup/steinbock/compare/v0.10.5...v0.10.6